+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1a7 (dev)
 
 - 新增：支持使用 OAS ParameterObject & MediaTypeObject examples。
+- 更新：model2schema 支持转换 JSONField。
 
 ## 0.1a6